Changed structure of the gradle build to be a multi module build.
The first module contains the production code for the annotation processor.

The second module contains a testing module to demontrate the annotation proccessor.

Still need to find a way to execute the annotation processor and generate the Builder
Java files without generating class files in the same folder as well